Add jenkins module package
Since the rtyler-jenkins module, which is a dependency of this module,
is not on the Forge I'm adding it directly into this module.

The rake setup task installs the module into the same directory this
module is installed into.  This satisfies the jenkins dependency in a
hackish, but effective way.