Rename to pe_jenkinsci adding pe_jenkinsci::facterjob
Without this patch, the scope of this module is too narrowly focused on
setting up the Facter spec tests.  This would have caused me to write
another module sitting in between the community jenkins module and the
pe_facterci module to setup the Jenkins service using PE.

This patch renames the module to pe_jenkinsci.  This module will be
responsible for the actual integration of Puppet Enterprise with Jenkins
CI.  The pe_jenkinsci::facterjob class will resolve the user story for
project empathy as part of this module.